# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberOS API - Barbershop SaaS platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Barbershop {
  id: number;
  name: string;
  slug: string;
  address: string;
  phone: string;
  description: string;
  logoUrl?: string | null;
  coverUrl?: string | null;
  openTime: string;
  closeTime: string;
  createdAt: string;
}

export interface CreateBarbershopBody {
  name: string;
  address: string;
  phone: string;
  description: string;
  openTime: string;
  closeTime: string;
}

export interface Barber {
  id: number;
  barbershopId: number;
  name: string;
  photoUrl?: string | null;
  bio: string;
  specialties: string;
  createdAt: string;
}

export interface CreateBarberBody {
  name: string;
  bio: string;
  specialties: string;
  photoUrl?: string | null;
}

export interface Service {
  id: number;
  barbershopId: number;
  name: string;
  description: string;
  durationMinutes: number;
  price: number;
  photoUrl?: string | null;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  description: string;
  durationMinutes: number;
  price: number;
  photoUrl?: string | null;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  email: string;
  phone: string;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export interface Appointment {
  id: number;
  barbershopId: number;
  barberId: number;
  clientId: number;
  serviceId: number;
  date: string;
  time: string;
  status: AppointmentStatus;
  notes?: string | null;
  barberName?: string | null;
  clientName?: string | null;
  serviceName?: string | null;
  servicePrice?: number | null;
  serviceDuration?: number | null;
  createdAt: string;
}

export interface CreateAppointmentBody {
  barberId: number;
  clientId: number;
  serviceId: number;
  date: string;
  time: string;
  notes?: string | null;
}

export type UpdateAppointmentBodyStatus =
  (typeof UpdateAppointmentBodyStatus)[keyof typeof UpdateAppointmentBodyStatus];

export const UpdateAppointmentBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;

export interface UpdateAppointmentBody {
  status?: UpdateAppointmentBodyStatus;
  date?: string;
  time?: string;
  notes?: string | null;
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  appointments: number;
}

export interface DashboardSummary {
  todayAppointments: number;
  monthRevenue: number;
  totalClients: number;
  completionRate: number;
  upcomingAppointments: Appointment[];
  revenueByDay: RevenueDataPoint[];
}

export type ListAppointmentsParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  status?: ListAppointmentsStatus;
};

export type ListAppointmentsStatus =
  (typeof ListAppointmentsStatus)[keyof typeof ListAppointmentsStatus];

export const ListAppointmentsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  completed: "completed",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;